requirements.txt
# Core dependencies
python-dotenv==1.0.0
openai==1.12.0
Pillow==10.2.0
opencv-python==4.9.0.80
requests==2.32.4

# GUI framework
customtkinter==5.2.2

# File handling and utilities
watchdog==4.0.0
pathlib2==2.3.7.post1

# Image and video processing
imageio==2.34.0
imageio-ffmpeg==0.4.9
numpy==1.26.4

# Configuration and logging
pyyaml==6.0.1
colorlog==6.8.2

# Testing (optional)
pytest==8.0.2
pytest-cov==4.1.0